feat: solana build api Tx input and wire-up in flip
