+++ packages/chain-adapters/src/solana/SolanaChainAdapter.ts
@@ -36,6 +36,7 @@ import { ErrorHandler } from '../error/ErrorHandler'
 import type {
   Account,
   BroadcastTransactionInput,
+  BuildSendApiTxInput,
   BuildSendTxInput,
   FeeDataEstimate,
   GetAddressInput,
@@ -203,14 +204,13 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.SolanaMainnet> 
     }
   }
 
-  async buildSendTransaction(input: BuildSendTxInput<KnownChainIds.SolanaMainnet>): Promise<{
+  async buildSendApiTransaction(input: BuildSendApiTxInput<KnownChainIds.SolanaMainnet>): Promise<{
     txToSign: SignTx<KnownChainIds.SolanaMainnet>
   }> {
     try {
-      const { accountNumber, to, chainSpecific, value } = input
+      const { from, accountNumber, to, chainSpecific, value } = input
       const { instructions = [], tokenId } = chainSpecific
 
-      const from = await this.getAddress(input)
       const { blockhash } = await this.connection.getLatestBlockhash()
 
       const computeUnitLimit = chainSpecific.computeUnitLimit
@@ -245,6 +245,19 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.SolanaMainnet> 
       }
 
       return { txToSign }
+    } catch (err) {
+      return ErrorHandler(err)
+    }
+  }
+
+  async buildSendTransaction(input: BuildSendTxInput<KnownChainIds.SolanaMainnet>): Promise<{
+    txToSign: SignTx<KnownChainIds.SolanaMainnet>
+  }> {
+    try {
+      const from = await this.getAddress(input)
+      const txToSign = await this.buildSendApiTransaction({ ...input, from })
+
+      return txToSign
     } catch (err) {
       return ErrorHandler(err)
     }
